+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Convert TODO to https://github.com/michael-cannon/remove-extra-media/issues
 * Convert to static class coding
 * Revise readme structure
+* Specify a “Text Domain” and “Domain Path”
 * transformmed > transformed
 
 ## 1.0.1